replace np.int with np.int32
ref: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ rvc_for_realtime.py
@@ -100,7 +100,7 @@ class RVC:
         ) + 1
         f0_mel[f0_mel <= 1] = 1
         f0_mel[f0_mel > 255] = 255
-        f0_coarse = np.rint(f0_mel).astype(np.int)
+        f0_coarse = np.rint(f0_mel).astype(np.int32)
         return f0_coarse, f0bak
 
     def get_f0(self, x, f0_up_key, n_cpu, method="harvest"):
